requirements.txt
# Core dependencies
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
chromadb>=0.4.22
pydantic>=2.5.0
requests>=2.31.0
numpy>=1.24.3
pandas>=2.0.3

# Audio processing (optional - uncomment if needed)
openai-whisper>=20231117
# torch>=2.1.0

# Google Gemini
google-generativeai>=0.3.2
httplib2>=0.22.0
google-api-python-client>=2.108.0
pyparsing>=3.1.1

# Optional: Hugging Face (only if needed)
# transformers>=4.35.0
# huggingface-hub>=0.19.0

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-cov>=4.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# bandit>=1.7.0
# safety>=2.0.0
# mypy>=1.0.0
